Split advanced_details into multiple files
